requirements.txt
werkzeug==2.0.1
Flask==1.1.2
polygon-api-client==0.2.11